requirements.txt
discord-py-interactions
discord.py==1.7.3
aiosqlite==0.17.0